requirements.txt
pandas
scikit-learn
joblib
google-cloud-storage
functions-framework
numpy